idempotent script

+++ scripts/init.sh
@@ -1,13 +1,29 @@
 #!/bin/sh
 
 if ! [ -f configs/.htpasswd ]
-  then echo 'admin:admin' >> configs/.htpasswd
+  then echo "==> Creating htpasswd"
+  echo 'admin:admin' >> configs/.htpasswd
 fi
 
-mkdir files
-mkdir downloads
-mkdir logs
+if ! [ -f configs/fileInfo.json ]
+  then echo "==> Creating fileInfo"
+  echo '{}' >> configs/fileInfo.json
+fi
 
-touch torrents.txt
+for dir in files downloads logs;
+do
+  if ! [ -d $dir ]
+    then echo "==> Creating $dir"
+    mkdir $dir
+  fi
+done
 
-cp configs/config.default configs/config.json
+if ! [ -f torrents.txt ]
+  then echo "==> Creating torrents"
+  touch torrents.txt
+fi
+
+if ! [ -f configs/config.json ]
+  then echo "==> Copying configs"
+  cp configs/config.default configs/config.json
+fi
